requirements.txt
Flask
opencv-python
ultralytics
numpy
pymongo[srv]
python-dotenv
cloudinary